Remove unused `use`
Since there are often duplicate branches, at least in my experience, -
due to remote and local being together, I've decided to keep the full
'path', like `origin/HEAD`, for example. So, `basename()` is no longer
needed.

It would be nice to have a version of this which doesn't include remote
branches. That kind of stuff will probably come at a later date,
though.