Add key `RegexStr` & `FullPath` (and more)
The user generated configuration file is now more verbose in its
comments. Some important details were previously left out, such as the
non-recursive nature of the default `./` value of `TodoStr`.

The REGEX of the `TodoStr` and `RegexStr` has been made more robust.

The `Char_Ofl` is now correctly included, even if `Pretty` is disabled.

Detection of valid keys is now much cleaner and probably optimized.

It's now possible to use `RegexStr` in the configuration file, to set a
permanent REGEX string to always filter out unwanted todo entries.

Stopped assigning the `TodoStr`, earlier on in the program, to the
value of `#TODO:`, as it's already done so in the configuration file.

Added a note in the usage output to prompt the user to adjust their
generated configuration file.