Finish (hopefully) Perl iteration of `lsbins`
Sadly, still using `man`, but it's possible to use `lsbins` without it
installed, but this will just omit the short description; still, it's
something.

Ideally, I'd like Perl to handle the desired part of what `man` does, -
but it would be an annoyance and a potentially-slower task, so I've
opted to stick with `man` itself to gather the short descriptions.

Bug fixes:

 * Fixed weird argument handling issues by using local `Arg` variable.

Additions:

 * Added Version comment for `List::MoreUtils` (for `uniq`).
 * Added usage information akin to other TFL programs.
 * Added `--extended-regexp|-E` & `--ignore-case|-i` MATCH options.
 * Added `--nodescribe|-N` option to omit short descriptions.
 * Added checks for user errors regarding above new options.
 * The old (temporary) approach has been replaced with good stuff.
 * Made the generation of executables more robust, with file testing.
 * Capitalizes the initial alphabetical character in the description.

Code changes:

 * Now using the latest (currently unpublished) TFL module.
 * Not using hashes, but instead arrays, which appear faster.