Add `gitsap` (Perl rewrite of shell `gitsa`)
Work in progress, but most of it is already done, so it should be
functional, including the configuration file.

The configuration file utility of `gitsap` borrows a lot of code and
structure from `tudu`, to save me a lot of hassle, plus I did a decent
job with it, so why not reuse it?